Rename files to use UpperCamelCaseCluster/Logic for software-diagnostics-server
