+++ src/app/clusters/software-diagnostics-server/app_config_dependent_sources.cmake
@@ -23,9 +23,9 @@ TARGET_SOURCES(
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/software-diagnostics-cluster.h"
-    "${CLUSTER_DIR}/software-diagnostics-logic.cpp"
-    "${CLUSTER_DIR}/software-diagnostics-logic.h"
+    "${CLUSTER_DIR}/SoftwareDiagnosticsCluster.h"
+    "${CLUSTER_DIR}/SoftwareDiagnosticsLogic.cpp"
+    "${CLUSTER_DIR}/SoftwareDiagnosticsLogic.h"
 )
 
 # These are the things that BUILD.gn dependencies would pull
